Fixing up quote for db
